Enhance build system for 32-bit support and improve CI configuration
- Add support for 32-bit builds with new build32.bat script and CMake presets
- Update CMakeLists.txt to handle architecture-specific output names
- Modify CI configuration to include matrix builds for both x64 and x86 architectures
- Refactor clang-tidy settings and caching for improved performance
- Introduce new game profile for IIDX 24 (SINOBUZ) with appropriate settings

+++ src/qpro_dll.cpp
@@ -125,7 +125,7 @@ std::string ReadCStr(const Image& img, uint64_t va) {
     std::string out;
     for (size_t i = 0; i <= kMaxName; ++i) {
         if ((size_t)off + i >= img.bytes.size()) return {};
-        uint8_t const c = img.bytes[off + i];
+        uint8_t const c = img.bytes[(size_t)off + i];
         if (c == 0) return out;
         if (c < 0x20 || c > 0x7E) return {};
         out.push_back((char)c);
